Refactor: Improve structure, error handling, and security
This commit introduces several improvements to the YRDSB Mark Fetcher application:

1.  **Code Refactoring:**
    *   I extracted core web scraping logic (mark fetching, advanced report fetching, and helper functions) from `app.py` and `appNoAI.py` into a new shared module: `yrdsb_scraper.py`. This significantly reduces code duplication and improves maintainability.
    *   Both Flask applications (`app.py` and `appNoAI.py`) now utilize this shared module.

2.  **Improved Error Handling:**
    *   I standardized error handling in `fetch_advanced_report` to return a user-friendly error message, similar to `fetch_marks`. The `advanced_report.html` template now displays these errors gracefully.
    *   I enhanced `yrdsb_scraper.py` to catch and handle network exceptions (`requests.exceptions.RequestException`) during HTTP requests.
    *   I added `try-except` blocks for critical HTML parsing operations in both `fetch_yrdsb_marks` and `fetch_yrdsb_advanced_report` to gracefully handle potential `AttributeError`, `IndexError`, or `TypeError` exceptions if the YRDSB portal's HTML structure changes. Errors are logged, and the scraper attempts to continue processing if possible.

3.  **Security Enhancements:**
    *   The Flask `app.secret_key` in `app.py` and `appNoAI.py` is now loaded from the `FLASK_SECRET_KEY` environment variable. A fallback key is provided for development convenience, but documentation strongly advises setting a secure environment variable.
    *   I verified that passwords are not stored in the `user_marks.json` file.

4.  **Unit Tests:**
    *   I added basic unit tests for helper functions (`parse_percentage`, `normalize_bgcolor`) in `yrdsb_scraper.py`.
    *   Test file: `tests/test_yrdsb_scraper.py`.

5.  **Documentation:**
    *   I updated `README.md` with instructions on how to set the `FLASK_SECRET_KEY` environment variable.
    *   I added a section to `README.md` explaining how to run the unit tests.

These changes make the application more robust, secure, and easier to maintain.
