+++ README.md
@@ -56,6 +56,36 @@ The application provides the following functionalities:
     ```
    If you do not have an api get one at https://aistudio.google.com/u/1/apikey free of charge (as of 01/14/25)
 
+### Security Configuration
+
+**Flask Secret Key**
+
+To ensure the security of user sessions, it's crucial to set a strong, random secret key for the Flask application.
+
+Set the `FLASK_SECRET_KEY` environment variable to a unique, random string. For example:
+
+On Linux/macOS:
+```bash
+export FLASK_SECRET_KEY='your_strong_random_secret_key_here'
+```
+
+On Windows (Command Prompt):
+```bash
+set FLASK_SECRET_KEY=your_strong_random_secret_key_here
+```
+
+On Windows (PowerShell):
+```bash
+$env:FLASK_SECRET_KEY='your_strong_random_secret_key_here'
+```
+
+You can generate a strong key using Python:
+```python
+import secrets
+secrets.token_hex(16)
+```
+The application includes a default fallback key for development if `FLASK_SECRET_KEY` is not set, but this fallback **should not** be used in any environment other than local development.
+
 # Running the Application
 Start the Flask server by running:
 
@@ -70,3 +100,16 @@ Or run the app without the AI Assistant with
   ```
 
 By default, the application will run on your local network and locally on http://{your_ip}:5001 and http://127.0.0.1:5001. Open this URL in your web browser to access the application.
+
+## Running Tests
+
+Basic unit tests are provided for some helper functions in `yrdsb_scraper.py`. To run them, navigate to the project's root directory and execute:
+
+```bash
+python -m unittest tests.test_yrdsb_scraper
+```
+Alternatively, you can run the test file directly:
+```bash
+python tests/test_yrdsb_scraper.py
+```
+The test file `tests/test_yrdsb_scraper.py` includes logic to adjust `sys.path` to ensure that the `yrdsb_scraper` module can be imported correctly when run directly.
